Move programs related hooks to program package
