+++ client/packages/programs/src/index.ts
@@ -1,3 +1,4 @@
 export * from './api';
 export * from './Components';
+export * from './hooks';
 export * from './JsonForms';
