+++ client/packages/system/src/Patient/PatientView/index.ts
@@ -1,8 +1,1 @@
 export { PatientView } from './PatientView';
-
-export enum PatientModal {
-  Prescription = 'PRESCRIPTION',
-  Program = 'PROGRAM',
-  ProgramSearch = 'PROGRAM_SEARCH',
-  Encounter = 'ENCOUNTER',
-}
